Guild Membership page now has Tabs for Members and Applicants
